classscatter {klaR}	R Documentation

Classification scatterplot matrix

Description

Function to plot a scatterplot matrix with a classification result.

Usage

classscatter(formula, data, method, col.correct = "black", 
    col.wrong = "red", gs = NULL, ...)

Arguments

formula
formula of the form groups ~ x1 + x2 + .... That is, the response is the grouping factor and the right hand side specifies the (non-factor) discriminators.

data
Data frame from which variables specified in formula are preferentially to be taken.

method
character, name of classification function (e.g. “lda”).

col.correct
color to use for correct classified objects.

col.wrong
color to use for missclassified objects.

gs
group symbol (plot character), must have the same length as the data. If NULL, as.character(groups) is the default.

...
further arguments passed to the underlying classification method or plot functions.

Value

The actual error rate.

Author(s)

Karsten Luebke, karsten.luebke@fom.de

See Also

plot

Examples

data(B3)
library(MASS)
classscatter(PHASEN ~ BSP91JW + EWAJW + LSTKJW, 
    data = B3, method = "lda")

[Package klaR version 1.7-4 Index]