cvtree {klaR}	R Documentation

Extracts variable cluster IDs

Description

Extracts cluster IDs for variables according to a dendrogram from object of class cvtree.

Usage

cvtree(object, k = 2, mincor = NULL, ...)

Arguments

object
Object of class corclust.

k
Number of clusters to be extracted from dendrogram.

mincor
Minimum within cluster correlation. Can be specified alternatively to k.

...
Currently not used.

Details

Like in corclust for correlation comparison numerics and factors are considered separately. For factors Cramer's V statistic is used.

Value

Object of class cvtree with elements:

cluster
Vector of cluster IDs.

correlations
Matrix of average within cluster correlations and average corrleation to all variables of the closest cluster as well as the ID of the closest cluster. For factor variables Cramer's V is computed.

Author(s)

Gero Szepannek

References

Roever, C. and Szepannek, G. (2005): Application of a genetic algorithm to variable selection in fuzzy clustering. In C. Weihs and W. Gaul (eds), Classification - The Ubiquitous Challenge, 674-681, Springer.

See Also

See also corclust, plot.corclust and hclust for details on the clustering algorithm.

Examples

    data(B3)
    ccres <- corclust(B3)
    plot(ccres)
    cvtree(ccres, k = 3)

[Package klaR version 1.7-4 Index]