drawparti {klaR}	R Documentation

Plotting the 2-d partitions of classification methods

Description

Plot showing the classification of observations based on classification methods (e.g. lda, qda) for two variables. Moreover, the classification borders are displayed and the apparent error rates are given in each title.

Usage

drawparti(grouping, x, y, method = "lda", prec = 100, xlab = NULL, 
    ylab = NULL, col.correct = "black", col.wrong = "red", 
    col.mean = "black", col.contour = "darkgrey", 
    gs = as.character(grouping), pch.mean = 19, cex.mean = 1.3, 
    print.err = 0.7, legend.err = FALSE, legend.bg = "white",
    imageplot = TRUE, image.colors = cm.colors(nc), 
    plot.control = list(), ...)        

Arguments

grouping
factor specifying the class for each observation.

x
first explanatory vector.

y
second explanatory vector.

method
the method the classification is based on, currently supported are: lda, qda, rpart, naiveBayes, rda, sknn and svmlight

.

prec
precision used to draw the classification borders (the higher the more precise; default: 100).

xlab
a title for the x axis.

ylab
a title for the y axis.

col.correct
color for correct classified objects.

col.wrong
color for wrong classified objects.

col.mean
color for class means (only for methods lda and qda).

col.contour
color of the contour lines (if imageplot = FALSE).

gs
group symbol (plot character), must have the same length as grouping.

pch.mean
plot character for class means (only for methods lda and qda).

cex.mean
character expansion for class means (only for methods lda and qda).

print.err
character expansion for text specifying the apparent error rate. If print.err = 0, nothing is printed.

legend.err
logical; whether to plot the apparent error rate above the plot (if FALSE), or into a legend into the upper right corner of the plot (if TRUE). This argument is ignored, if print.err = 0, i.e. if no error rate is printed.

legend.bg
Backgound colour to use for the legend.

imageplot
logical; whether to use an image plot or contour lines.

image.colors
colors used for the imageplot, if TRUE.

plot.control
A list containing further arguments passed to the underlying plot functions.

...
Further arguments passed to the classification method.

Author(s)

Karsten Luebke, karsten.luebke@fom.de, Uwe Ligges, Irina Czogiel

See Also

partimat

[Package klaR version 1.7-4 Index]