plot.woe {klaR}	R Documentation

Plot information values

Description

Barplot of information values to compare dicriminator of the transformed variables.

Usage

## S3 method for class 'woe'
plot(x, type = c("IV", "woes"), ...)

Arguments

x
An object of class woe.

type
Character to specify the plot type, see below. Either "IV" (default) or "woes".

...
Further arguments to be passed to the barplot function.

Details

For type=="IV" a barplot of information values for all transformed variables. A thumb rule of interpretation is that Values above 0.3 are considered as strongly discrimative where values below 0.02 are considered to characterize unpredictive variables. For type=="woes" for each variable the relative frequencies of all transformed levels are plotted.

Value

No value is returned.

Author(s)

Gero Szepannek

References

Good, I. (1950): Probability and the Weighting of Evidences. Charles Griffin, London.

Kullback, S. (1959): Information Theory and Statistics. Wiley, New York.

See Also

woe, predict.woe

Examples

# see examples in ?woe

[Package klaR version 1.7-4 Index]