predict.loclda {klaR}	R Documentation

Localized Linear Discriminant Analysis (LocLDA)

Description

Classifies new observations using parameters determined by the loclda-function.

Usage

## S3 method for class 'loclda'
predict(object, newdata, ...)

Arguments

object
Object of class loclda.

newdata
Data frame of cases to be classified.

...
Further arguments are ignored.

Value

A list with components:

class
Vector (of class factor) of classifications.

posterior
Posterior probabilities for the classes. For details of computation see loclda (+ normalization so posterior-values add up to 1 for each observation).

all.zero
Vector (of class integer) indicating for which rows of newdata all corresponding posterior-values are < 1e-150 before normalization. Those observations are assigned to the class to whose (locally weighted) centroid they have the lowest euclidian distance.

Author(s)

Marc Zentgraf (marc-zentgraf@gmx.de) and Karsten Luebke (karsten.luebke@fom.de)

See Also

loclda

Examples

data(B3)
x <- loclda(PHASEN ~ ., data = B3, subset = 1:80)
predict(x, B3[-(1:80),])

[Package klaR version 1.7-4 Index]