rerange {klaR}	R Documentation

Linear transformation of data

Description

The function performs a linear transformation of the data, such that afterwards range(data)=c(theMin,theMax).

Usage

rerange(data, min.goal = 0, max.goal = 1, min.data =
    min(data), max.data = max(data), center = NA)

Arguments

data
vector with the data to transform

min.goal
new minimum value

max.goal
new maximum value

min.data
old minimum value

max.data
old maximum value

center
which old value should become the new center (max.goal + min.goal) / 2)

Value

vector with the transformed data

Author(s)

Dominik Reusser

Examples

   klaR:::rerange(data=1:20)
   klaR:::rerange(data=1:30, center=5)

[Package klaR version 1.7-4 Index]