triframe {klaR}	R Documentation

Barycentric plots

Description

Function to add a frame to an existing (barycentric) plot.

Usage

triframe(label = 1:3, label.col = 1, cex = 1, ...)

Arguments

label
labels for the three corners of the plot.

label.col
text color for labels.

cex
Magnification factor for label text relative to the default.

...
Further graphical parameters passed to trilines.

Author(s)

Christian Röver, roever@statistik.tu-dortmund.de

See Also

triplot, trilines, trigrid, centerlines

Examples

triplot(grid = TRUE, frame = FALSE)     # plot without frame 
some.triangle <- rbind(c(0, 0.65, 0.35), c(0.53, 0.47, 0), 
                       c(0.72, 0, 0.28))[c(1:3, 1), ]
trilines(some.triangle, col = "red", pch = 16, type = "b")
triframe(label = c("left", "top", "right"), col = "blue", 
         label.col = "green3")          # frame on top of points 

[Package klaR version 1.7-4 Index]