triperplines {klaR}	R Documentation

Barycentric plots

Description

Function to add a point and the corresponding perpendicular lines to all three sides to an existing (barycentric) plot.

Usage

triperplines(x, y = NULL, z = NULL, lcol = "red", pch = 17, ...)

Arguments

x
fraction of first component OR 3-element vector (for all three components, omitting y and z).

y
(optional) fraction of second component.

z
(optional) fraction of third component.

lcol
line color

pch
plotting character. pch = 0 for no point

...
Further graphical parameters (see points, lines and par).

Details

Adds a (single!) point and lines to an existing plot (generated by triplot). The lines originate from the point and run (perpendicular) towards all three sides. The lengths (and proportions) of these lines are identical to those of x, y and z.

Value

a 2-column-matrix containing plot coordinates.

Author(s)

Christian Röver, roever@statistik.tu-dortmund.de

See Also

triplot, tripoints, trilines, tritrafo

Examples

triplot()  # empty plot 
triperplines(1/2, 1/3, 1/6)

[Package klaR version 1.7-4 Index]