tritrafo {klaR}	R Documentation

Barycentric plots

Description

Function to carry out the transformation into 2D space for triplot, trilines etc.

Usage

tritrafo(x, y = NULL, z = NULL, check = TRUE, tolerance = 0.0001)

Arguments

x
Vector of fractions of first component OR 3-column matrix containing all three components (omitting y and z) OR 3-element vector (for all three components, omitting y and z).

y
(optional) vector of fractions of second component.

z
(optional) vector of fractions of third component.

check
if TRUE, it is checked whether x+y+z=1 and x,y,z>=0 for all cases.

tolerance
tolerance for above sum check.

Details

Projects the mixture given by x, y, and z with x, y, z between one and zero and x+y+z=1 into a two-dimensional space.

For further details see triplot.

Value

A matrix with two columns corresponding to the two dimensions.

Author(s)

Christian Röver, roever@statistik.tu-dortmund.de

See Also

triplot, tripoints, trilines, trigrid

Examples

tritrafo(0.1, 0.2, 0.7)
tritrafo(0.1, 0.2, 0.6) # warning 

triplot()
points(tritrafo(0.1, 0.2, 0.7), col="red")
tripoints(0.1, 0.2, 0.7, col="green")  # the same

tritrafo(c(0.1,0.2), c(0.3,0.4), c(0.6,0.4))
tritrafo(diag(3))

point <- c(0.25,0.6,0.15)
triplot(point, pch=16)
text(tritrafo(point), "(0.25, 0.60, 0.15)", adj=c(0.5,2)) # add a label 

[Package klaR version 1.7-4 Index]