ucpm {klaR}	R Documentation

Uschi's classification performance measures

Description

Function to calculate the Correctness Rate, the Accuracy, the Ability to Seperate and the Confidence of a classification rule.

Usage

ucpm(m, tc, ec = NULL)

Arguments

m
matrix of (scaled) membership values

tc
vector of true classes

ec
vector of estimated classes (only required if scaled membership values are used)

Details

The correctness rate is the estimator for the correctness of a classification rule (1-error rate).

The accuracy is based on the euclidean distances between (scaled) membership vectors and the vectors representing the true class corner. These distances are standardized so that a measure of 1 is achieved if all vectors lie in the correct corners and 0 if they all lie in the center.

Analougously, the ability to seperate is based on the distances between (scaled) membership vectors and the vector representing the corresponding assigned class corner.

The confidence is the mean of the membership values of the assigned classes.

Value

A list with elements:

CR
Correctness Rate

AC
Accuracy

AS
Ability to Seperate

CF
Confidence

CFvec
Confidence for each (true) class

Author(s)

Karsten Luebke, karsten.luebke@fom.de

References

Garczarek, Ursula Maria (2002): Classification rules in standardized partition spaces. Dissertation, University of Dortmund. URL http://hdl.handle.net/2003/2789

Examples

library(MASS)
data(iris)
ucpm(predict(lda(Species ~ ., data = iris))$posterior, iris$Species)

[Package klaR version 1.7-4 Index]