xtractvars {klaR}	R Documentation

Variable clustering based variable selection

Description

Applies variable selection to data based on variable clusterings as resulting from corclust or CLV.

Usage

xtractvars(object, data, thres = 0.5)

Arguments

object
Object of class cvtree applied to a corclust object or the summary() of a clv object as created by CLV.

data
Data where variables are to be selected. Coloumn names must be identical to those used in corclust model.

thres
Maximum accepted average within cluster correlation for selection of a variable.

Details

Of each cluster the first variable is selected as well as all other variables with an average within cluster correlation below thres.

Value

The data is returned where unselected coloumns are removed.

Author(s)

Gero Szepannek

References

Roever, C. and Szepannek, G. (2005): Application of a genetic algorithm to variable selection in fuzzy clustering. In C. Weihs and W. Gaul (eds), Classification - The Ubiquitous Challenge, 674-681, Springer.

See Also

See also corclust, cvtree and CLV.

Examples

    data(B3)
    ccres <- corclust(B3)
    plot(ccres)
    cvtres <- cvtree(ccres, k = 3)
    newdata <- xtractvars(cvtres, B3, thres = 0.5) 

[Package klaR version 1.7-4 Index]