TopoS {klaR}	R Documentation

Computation of criterion S of a visualization

Description

TopoS computes one version of the criterion stress as i.e. known from sammon for a given visualization.

Usage

TopoS(EV.dist, Cells.dist)

Arguments

EV.dist
a symmetric distance matrix consisting of distances in the origin space (dx)

Cells.dist
a symmetric distance matrix consisting of distances in the target space (dy)

Details

TopoS first performs a homogeneous linear regression where dx is predicted by dy. Then the residual sum of squares of this regression is computed and divided by the sum of squared dx. Finally one minus the square root of the latter result defines S.

Value

TopoS returns a numeric between zero and one which is the criterion S of the given visualization.

Author(s)

Nils Raabe

See Also

EDAM

Examples

# Compute S for the MDS visualization of the german business data
data(B3)
plot(cmdscale(dist(B3[, 2:14])), col = rainbow(4)[B3[, 1]], pch = 16)
TopoS(dist(B3[, 2:14]), dist(cmdscale(dist(B3[, 2:14]))))

[Package klaR version 0.6-12 Index]