b.scal {klaR}	R Documentation

Calculation of beta scaling parameters

Description

Calculates the scaling parameter for betascale.

Usage

b.scal(member, grouping, dis = FALSE, eps = 1e-04)

Arguments

member
Membership values of an argmax classification method. Eg. posterior probabilities of lda. Row-wise values must sum up to 1 and must be in the interval [0,1].

grouping
Class vector.

dis
Logical, whether to optimize the dispersion parameter in pbeta.

eps
Minimum variation of membership values. If variance is smaller than eps, the values are treated as one point.

Details

With betascale and b.scal, membership values of an argmax classifier are scaled in such a way, that the mean membership value of those values which are assigned to each class reflect the mean correctness rate of that values. This is done via qbeta and pbeta with the appropriate shape parameters. If dis is TRUE, it is tried that the variation of membership values is optimal for the accuracy relative to the correctness rate. If the variation of the membership values is less than eps, they are treated as one point and shifted towards the correctness rate.

Value

A list containing

model
Estimated parameters for betascale.

eps
Value of eps from the call.

member
Scaled membership values.

Author(s)

Karsten Luebke (karsten.luebke@fom.de), Uwe Ligges

References

Garczarek, Ursula Maria (2002): Classification rules in standardized partition spaces. Dissertation, University of Dortmund. URL http://hdl.handle.net/2003/2789

See Also

betascale, e.scal

Examples

library(MASS)
data(B3)
pB3 <- predict(lda(PHASEN ~ ., data = B3))$posterior
pbB3 <- b.scal(pB3, B3$PHASEN, dis = TRUE)
ucpm(pB3, B3$PHASEN)
ucpm(pbB3$member, B3$PHASEN)

[Package klaR version 0.6-12 Index]