benchB3 {klaR}	R Documentation

Benchmarking on B3 data

Description

Evaluates the performance of a classification method on the B3 data.

Usage

benchB3(method, prior = rep(1/4, 4), sv = "4", scale = FALSE, ...)

Arguments

method
classification method to use

prior
prior probabilities of classes

sv
class of the start of a business cycle

scale
logical, whether to use scale first

...
furhter arguments passed to method

Details

The performance of classification methods on cyclic data can be measured by a special form of cross-validation: Leave-One-Cycle-Out. That means that a complete cycle is used as test data and the others are used as training data. This is repeated for all complete cycles in the data.

Value

A list with elements

MODEL
list with the model returned by method of the training data

error
vector of test error rates in cycles

l1co.error
leave-one-cycle-out error rate

Author(s)

Karsten Luebke, karsten.luebke@fom.de

See Also

B3

Examples

perLDA <- benchB3("lda")
## Not run: 
## due to parameter optimization rda takes a while 
perRDA <- benchB3("rda")
library(rpart)
## rpart will not work with prior argument:
perRpart <- benchB3("rpart", prior = NULL)

## End(Not run)

[Package klaR version 0.6-12 Index]