calc.trans {klaR}	R Documentation

Calculation of transition probabilities

Description

Function to estimate the probabilities of a time series to stay or change the state.

Usage

calc.trans(x)

Arguments

x
(factor) vector of states

Details

To estimate the transition probabilities the empirical frequencies are counted.

Value

The transition probabilities matrix. x[i,j] is the probability to change from state i to state j.

Author(s)

Karsten Luebke, karsten.luebke@fom.de

Examples

data(B3)
calc.trans(B3$PHASEN)

[Package klaR version 0.6-12 Index]