centerlines {klaR}	R Documentation

Lines from classborders to the center

Description

Function which constructs the lines from the borders between two classes to the center. To be used in connection with triplot and quadplot.

Usage

centerlines(n)

Arguments

n
number of classes. Meaningful are 3 or 4.

Value

a matrix with n-columns.

Author(s)

Karsten Luebke, karsten.luebke@fom.de

See Also

triplot, quadplot

Examples

centerlines(3)
centerlines(4)

[Package klaR version 0.6-12 Index]