cond.index {klaR}	R Documentation

Calculation of Condition Indices for Linear Regression

Description

Diagnosis of collinearity in X

Usage

cond.index(formula, data, ...)

Arguments

formula
formula of the form ‘groups ~ x1 + x2 + ...’

data
data frame (or matrix) containing the explanatory variables

...
further arguments to be passed to lm

Details

Collinearities can inflate the variance of the estimated regression coefficients and numerical stability. The condition indices are calculated by the eigenvalues of the crossproduct matrix of the scaled but uncentered explanatory variables. Indices > 30 may indicate collinearity.

Value

A vector of the condition indices.

Author(s)

Andrea Preusser, Karsten Luebke (karsten.luebke@fom.de)

References

Belsley, D. , Kuh, E. and Welsch, R. E. (1979), Regression Diagnostics: Identifying Influential Data and Sources of Collinearity, John Wiley (New York)

See Also

stepclass, manova

Examples

data(Boston)
condition_medv <- cond.index(medv ~ ., data = Boston)
condition_medv

[Package klaR version 0.6-12 Index]