meclight.default {klaR}	R Documentation

Minimal Error Classification

Description

Computer intensive method for linear dimension reduction that minimizes the classification error directly.

Usage

meclight(x, ...)

## Default S3 method:
meclight(x, grouping, r = 1, fold = 10, ...)
## S3 method for class 'formula'
meclight(formula, data = NULL, ..., subset, na.action = na.fail)
## S3 method for class 'data.frame'
meclight(x, ...)
## S3 method for class 'matrix'
meclight(x, grouping, ..., subset, na.action = na.fail)

Arguments

x
(required if no formula is given as the principal argument.) A matrix or data frame containing the explanatory variables.

grouping
(required if no formula principal argument is given.) A factor specifying the class for each observation.

r
Dimension of projected subspace.

fold
Number of Bootstrap samples.

formula
A formula of the form groups ~ x1 + x2 + .... That is, the response is the grouping factor and the right hand side specifies the (non-factor) discriminators.

data
Data frame from which variables specified in formula are preferentially to be taken.

subset
An index vector specifying the cases to be used in the training sample. (NOTE: If given, this argument must be named.)

na.action
A function to specify the action to be taken if NAs are found. The default action is for the procedure to fail. An alternative is na.omit, which leads to rejection of cases with missing values on any required variable. (NOTE: If given, this argument must be named.)

...
Further arguments passed to lda.

Details

Computer intensive method for linear dimension reduction that minimizes the classification error in the projected subspace directly. Classification is done by lda. In contrast to the reference function minimization is done by Nelder-Mead in optim.

Value

method.model
An object of class ‘lda’.

Proj.matrix
Projection matrix.

B.error
Estimated bootstrap error rate.

B.impro
Improvement in lda error rate.

Author(s)

Maria Eveslage, Karsten Luebke, karsten.luebke@fom.de

References

Roehl, M.C., Weihs, C., and Theis, W. (2002): Direct Minimization in Multivariate Classification. Computational Statistics, 17, 29-46.

See Also

predict.meclight

Examples

data(iris)
meclight.obj <- meclight(Species ~ ., data = iris)
meclight.obj

[Package klaR version 0.6-12 Index]