predict.locpvs {klaR}	R Documentation

predict method for locpvs objects

Description

Prediction of class membership and posterior probabilities in local models using pairwise variable selection.

Usage

## S3 method for class 'locpvs'
predict(object,newdata, quick = FALSE, return.subclass.prediction = TRUE, ...)

Arguments

object
an object of class ‘locpvs’, as that created by the function “locpvs”

newdata
a data frame or matrix containing new data. If not given the same datas as used for training the ‘pvs’-model are used.

quick
indicator (logical), whether a quick, but less accurate computation of posterior probabalities should be used or not.

return.subclass.prediction
indicator (logical), whether the returned object includes posterior probabilities for each date in each subclass

...
Further arguments are passed to underlying predict calls.

Details

Posterior probabilities are predicted as if object is a standard ‘pvs’-model with the subclasses as classes. Then the posterior probabalities are summed over all subclasses for each class. The class with the highest value becomes the prediction.

If “quick=FALSE” the posterior probabilites for each case are computed using the pairwise coupling algorithm presented by Hastie, Tibshirani (1998). If “quick=FALSE” a much quicker solution is used, which leads to less accurate posterior probabalities. In almost all cases it doesn't has a negative effect on the classification result.

Value

a list with components:

class
the predicted (upper) classes

posterior
posterior probabilities for the (upper) classes

subclass.posteriors
(only if “return.subclass.prediction=TRUE”. A matrix containing posterior probabalities for the subclasses.

Author(s)

Gero Szepannek, szepannek@statistik.tu-dortmund.de, Christian Neumann

References

Szepannek, G. and Weihs, C. (2006) Local Modelling in Classification on Different Feature Subspaces. In Advances in Data Mining., ed Perner, P., LNAI 4065, pp. 226-234. Springer, Heidelberg.

See Also

locpvs for learning ‘locpvs’-models and examples for applying this predict method, pvs for pairwise variable selection without modeling subclasses, predict.pvs for predicting ‘pvs’-models

[Package klaR version 0.6-12 Index]