predict.meclight {klaR}	R Documentation

Prediction of Minimal Error Classification

Description

Classify multivariate observations in conjunction with meclight and lda.

Usage

## S3 method for class 'meclight'
predict(object, newdata,...)

Arguments

object
Object of class meclight.

newdata
Data frame of cases to be classified or, if object has a formula, a data frame with columns of the same names as the variables used. A vector will be interpreted as a row vector.

...
currently ignored

Details

Classify multivariate observations in conjunction with meclight and lda.

Value

class
The estimated class (factor).

posterior
Posterior probabilities for the classes.

Author(s)

Karsten Luebke, karsten.luebke@fom.de

References

Roehl, M.C., Weihs, C., and Theis, W. (2002): Direct Minimization in Multivariate Classification. Computational Statistics, 17, 29-46.

See Also

meclight

Examples

data(iris)
meclight.obj <- meclight(Species ~ ., data = iris)
predict(meclight.obj, iris)

[Package klaR version 0.6-12 Index]