predict.woe {klaR}	R Documentation

Weights of evidence

Description

Applies weight of evidence transform of factor variables for binary classification based on a model of class woe.

Usage

## S3 method for class 'woe'
predict(object, newdata, replace = TRUE, ...)

Arguments

object
Object resulting from a call of woe.

newdata
A matrix or data frame where WOE transform should be applied of the same dimension as the data used for training the woe object.

replace
Logical flag specifying whether the original factor variables should be kept in the output.

...
Currently not used.

Value

Data frame including the transformed numeric woe variables.

Author(s)

Gero Szepannek

References

Good, I. (1950): Probability and the Weighting of Evidences. Charles Griffin, London.

See Also

woe, plot.woe

Examples

# see examples in ?woe

[Package klaR version 0.6-12 Index]