quadtrafo {klaR}	R Documentation

Transforming of 4 dimensional values in a barycentric coordinate system.

Description

Transforming of 4 dimensional values in a barycentric coordinate system.

Usage

quadtrafo(e, f = NULL, g = NULL, h = NULL)
        
quadlines(e, f = NULL, g = NULL, h = NULL, sp, ...)

quadpoints(e, f = NULL, g = NULL, h = NULL, sp, ...)

Arguments

e
either a matrix with 4 columns represanting the membership values or a vector with the membership values of the first class

f
vector with the membership values of the second class

g
vector with the membership values of the third class

h
vector with the membership values of the forth class

sp
scatterplot3d object to which points or lines should be plotted

...
further arguments passed to the underlyind plot functions

Value

For quadtrafo the 3 dimensional values in the barycentrix coordinate system.

Author(s)

Karsten Luebke, karsten.luebke@fom.de

See Also

triplot, quadplot

Examples

library(MASS)
data(B3)
posterior <- predict(lda(PHASEN ~ ., data = B3))$post
quadtrafo(posterior)

[Package klaR version 0.6-12 Index]