tripoints {klaR}	R Documentation

Barycentric plots

Description

Function to add points or lines to an existing (barycentric) plot.

Usage

tripoints(x, y = NULL, z = NULL, ...)
trilines(x, y = NULL, z = NULL, ...)

Arguments

x
Vector of fractions of first component OR 3-column matrix containing all three components (omitting y and z) OR 3-element vector (for all three components, omitting y and z).

y
(optional) vector of fractions of second component.

z
(optional) vector of fractions of third component.

...
Further graphical parameters (see points and par).

Details

Adds points or lines to an existing plot (generated by triplot).

Author(s)

Christian Röver, roever@statistik.tu-dortmund.de

See Also

points, lines, triplot, tritrafo, centerlines

Examples

triplot()  # empty plot 
tripoints(0.1, 0.2, 0.7)                        # a point 
tripoints(c(0.2, 0.6), c(0.3, 0.3), c(0.5, 0.1),
    pch = c(2, 6))                              # two points
trilines(c(0.1, 0.6), c(0.2, 0.3), c(0.7, 0.1), 
    col = "blue", lty = "dotted")               # a line 

trilines(centerlines(3))

[Package klaR version 0.6-12 Index]